refactor(web): lift assistant-context + active-assistant-gate out of chat domain
These are route infrastructure (outlet context type + lifecycle gate),
not chat behavior — every page route consumes them. Moving them to
`src/components/layout/` (next to `root-layout.tsx`) clears the
last reason for five non-chat domain pages to reach into chat:

- contacts/connect-page, contacts/contacts-page
- intelligence/identity-page, intelligence/skills-page
- workspace/workspace-page

Five allowlist entries drop to zero deps and are removed; library-page
drops its "chat" dep (kept "intelligence" for an unrelated import).

Pure refactor — no behavior change.
