+++ apps/web/src/components/layout/active-assistant-gate.tsx
@@ -6,7 +6,7 @@ import { Typography } from "@vellum/design-library";
 import {
   useAssistantContext,
   type AssistantContextValue,
-} from "@/domains/chat/assistant-context.js";
+} from "@/components/layout/assistant-context.js";
 
 /**
  * Narrowed outlet context for routes mounted inside `ActiveAssistantGate`.
