+++ apps/web/src/components/layout/assistant-context.ts
