+++ apps/web/src/domains/intelligence/skills-page.tsx
@@ -1,6 +1,6 @@
 import { useSearchParams } from "react-router";
 
-import { useActiveAssistantContext } from "@/domains/chat/active-assistant-gate.js";
+import { useActiveAssistantContext } from "@/components/layout/active-assistant-gate.js";
 import { SkillsTab } from "@/domains/intelligence/components/skills/skills-tab.js";
 
 export function SkillsPage() {
