+++ apps/web/src/domains/library/library-page.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { useNavigate } from "react-router";
 
-import { useActiveAssistantContext } from "@/domains/chat/active-assistant-gate.js";
+import { useActiveAssistantContext } from "@/components/layout/active-assistant-gate.js";
 import { LibraryView } from "@/domains/intelligence/components/apps/library-view.js";
 import { routes } from "@/utils/routes.js";
